Refactor : 모듈화 및 폴더구조 변경, 데이터 관리 레이어 추가(미션2 대비)
