+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>이벤트 - TODOS</title>
     <link rel="stylesheet" href="./src/css/style.css" />
-    <script type="module" src="./src/todo/todoapp.js" defer="defer"></script>
+    <script type="module" src="./src/js/App.js" defer="defer"></script>
   </head>
   <body>
     <div class="todoapp">
