+++ src/js/util/domSelection.js
