--- src/todo/todoapp.js
@@ -1,203 +1,0 @@
-import { $, $$ } from "../util/domSelection.js";
-import { DAO } from "../datastore/datastore.js";
-class TodoApp {
-  constructor(todoItemArray) {
-    this.todoItemArray = todoItemArray;
-
-    const newTodoInput = $("#new-todo-title");
-    newTodoInput.addEventListener("keydown", (e) => {
-      if (e.key == "Enter") {
-        this.addItem(newTodoInput.value);
-        newTodoInput.value = "";
-      }
-    });
-  }
-  loadData() {
-    const localStoredArray = DAO.loadData();
-    localStoredArray.forEach((item) => {
-      this.todoItemArray.push(new TodoItem(item.data, item.state));
-    });
-    this.setState();
-  }
-  addItem(data) {
-    if (!data || data.trim() == "") return;
-    this.todoItemArray.push(new TodoItem(data, TodoItem.ACTIVE));
-    this.setState();
-  }
-  deleteItem(index) {
-    this.todoItemArray.splice(index, 1);
-    this.setState();
-  }
-  updateItem(index, data) {
-    this.todoItemArray[index].data = data;
-    this.setState();
-  }
-  updateItemState(index, isCompleted) {
-    if (isCompleted) {
-      this.todoItemArray[index].state = TodoItem.COMPLETED;
-    } else {
-      this.todoItemArray[index].state = TodoItem.ACTIVE;
-    }
-    this.setState();
-  }
-  setState() {
-    if (this.todoList) {
-      this.todoList.setState(this.todoItemArray);
-    }
-    if (this.todoStatusContainer) {
-      this.todoStatusContainer.setState();
-    }
-    DAO.saveData(this.todoItemArray);
-  }
-}
-
-class TodoItem {
-  static ACTIVE = "active";
-  static COMPLETED = "completed";
-
-  constructor(data, state) {
-    if (state != TodoItem.ACTIVE && state != TodoItem.COMPLETED) {
-      alert("item state error");
-      return;
-    }
-    this.data = data;
-    this.state = state;
-  }
-  toString() {
-    return JSON.stringify(this);
-  }
-}
-
-class TodoList {
-  constructor(todoApp) {
-    this.todoApp = todoApp;
-
-    const list = $("#todo-list");
-    list.addEventListener("click", function (e) {
-      if (e.target && e.target.className == "toggle") {
-        const targetLi = e.target.closest("li");
-        todoApp.updateItemState(
-          targetLi.dataset.index,
-          targetLi.classList.toggle(TodoItem.COMPLETED)
-        );
-        targetLi.classList.toggle(TodoItem.ACTIVE);
-      }
-    });
-    list.addEventListener("click", function (e) {
-      if (e.target && e.target.className == "destroy") {
-        const targetLi = e.target.closest("li");
-        targetLi.outerHTML = "";
-        todoApp.deleteItem(targetLi.dataset.index);
-      }
-    });
-    list.addEventListener("dblclick", function (e) {
-      if (e.target && e.target.nodeName == "LABEL") {
-        const targetLi = e.target.closest("li");
-        targetLi.classList.add("editing");
-      }
-    });
-    list.addEventListener("keydown", function (e) {
-      if (e.target && e.target.nodeName == "INPUT") {
-        const targetLi = e.target.closest("li");
-        if (e.key == "Escape") {
-          targetLi.classList.remove("editing");
-        } else if (e.key == "Enter") {
-          todoApp.updateItem(targetLi.dataset.index, e.target.value);
-          targetLi.classList.remove("editing");
-        }
-      }
-    });
-  }
-  setState(todoItemArray) {
-    const list = $("#todo-list");
-    list.innerHTML = "";
-    let li;
-    let index = 0;
-    todoItemArray.forEach((item) => {
-      li = document.createElement("li");
-
-      const indexAttribute = document.createAttribute("data-index");
-      indexAttribute.value = index;
-      index = index + 1;
-      li.setAttributeNode(indexAttribute);
-
-      if (item.state == TodoItem.COMPLETED) {
-        const liClass = document.createAttribute("class");
-        liClass.value = TodoItem.COMPLETED;
-        li.setAttributeNode(liClass);
-      } else if (item.state == TodoItem.ACTIVE) {
-        const liClass = document.createAttribute("class");
-        liClass.value = TodoItem.ACTIVE;
-        li.setAttributeNode(liClass);
-      }
-      li.innerHTML = `
-            <div class="view">
-                <input class="toggle" type="checkbox" ${
-                  item.state == TodoItem.COMPLETED ? "checked" : ""
-                }/>
-                <label class="label">${item.data}</label>
-                <button class="destroy"></button>
-            </div>
-            <input class="edit" value="${item.data}" />
-            `;
-      list.appendChild(li);
-    });
-  }
-}
-class TodoStatusContainer {
-  constructor() {
-    const filters = {
-      activeFilter: {
-        button: $(".count-container ." + TodoItem.ACTIVE),
-        name: TodoItem.ACTIVE,
-      },
-      completedFilter: {
-        button: $(".count-container ." + TodoItem.COMPLETED),
-        name: TodoItem.COMPLETED,
-      },
-      allFilter: { button: $(".count-container .all"), name: "all" },
-    };
-
-    const hideSelectedFilter = () =>
-      $$(".count-container a").forEach((button) =>
-        button.classList.remove("selected")
-      );
-    const showSelectedFilter = (selectedFilter) => {
-      const todoListLi = $$("#todo-list li");
-      selectedFilter.button.classList.add("selected");
-      if (selectedFilter.name == "all") {
-        todoListLi.forEach((li) => (li.style.display = ""));
-        return;
-      }
-      todoListLi.forEach((li) => {
-        if (!li.classList.contains(selectedFilter.name))
-          li.style.display = "none";
-        else li.style.display = "";
-      });
-    };
-
-    Object.entries(filters).forEach(([key, filter]) => {
-      filter.button.addEventListener("click", (e) => {
-        e.preventDefault();
-        hideSelectedFilter();
-        showSelectedFilter(filter);
-        this.setState();
-      });
-    });
-  }
-  setState() {
-    let count = 0;
-    $$("#todo-list li").forEach((li) => {
-      if (li.style.display != "none") count = count + 1;
-    });
-    $(".todo-count strong").textContent = count;
-  }
-}
-
-const todoApp = new TodoApp([]);
-const todoList = new TodoList(todoApp);
-const todoStatusContainer = new TodoStatusContainer();
-
-todoApp.todoList = todoList;
-todoApp.todoStatusContainer = todoStatusContainer;
-todoApp.loadData();
